--- utils/tux.py
@@ -1,6 +1,0 @@
-from ..constants import tuxURL
-import json
-import urllib.parse
-import hmac
-import hashlib
-
